Extract DependencyResolutionNode to separate file
